Add launch options method of starting to README
This method, apart from looking cleaner, allows the multiplayer invitation feature to work… maybe. When I tried it, the game failed to connect but I'm not sure whether that was due to the conversion, the fact that we were on the same LAN, or some other issue. There have been reports of difficulties with the multiplayer lobby on every platform.